refactor(agglayer): share the bridge setup helper across test suites
Extract the duplicated setup_bridge helpers from remove_ger.rs and
rbac_rotation.rs into test_utils (BridgeSetup + setup_bridge), seeding
a chain-resident admin wallet so both suites - and future ones - can
use the same fixture.

Addresses review feedback on re-using test helpers.
